csharp/HangMan/Gallow.cs
using System;
using System.Collections;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace HangMan {
  public class Gallow {
    readonly private static List<string> stage = new () {
      "  +---+\n" +
      "  |   |\n" +
      "      |\n" +
      "      |\n" +
      "      |\n" +
      "      |\n" +
      "=========\n",
      "  +---+\n" +
      "  |   |\n" +
      "  O   |\n" +
      "      |\n" +
      "      |\n" +
      "      |\n" +
      "=========\n",
      "  +---+\n" +
      "  |   |\n" +
      "  O   |\n" +
      "  |   |\n" +
      "      |\n" +
      "      |\n" +
      "=========\n",
      "  +---+\n" +
      "  |   |\n" +
      "  O   |\n" +
      " /|   |\n" +
      "      |\n" +
      "      |\n" +
      "=========\n",
      "  +---+\n" +
      "  |   |\n" +
      "  O   |\n" +
      " /|\\  |\n" +
      "      |\n" +
      "      |\n" +
      "=========\n",
      "  +---+\n" +
      "  |   |\n" +
      "  O   |\n" +
      " /|\\  |\n" +
      " /    |\n" +
      "      |\n" +
      "=========\n",
      "  +---+\n" +
      "  |   |\n" +
      "  O   |\n" +
      " /|\\  |\n" +
      " / \\  |\n" +
      "      |\n" +
      "=========\n"
    };
    public static string ReturnStage(int i) {
      try {
        return stage[6 - i];
      }
      catch (Exception e) {
        Console.WriteLine(e.Message);
        throw new Exception("index parameter is out of range.", e);
      }
    }
  }
}

